fix(ci): split collect_modules.rs under the 2000-line file cap
lint on main has a second failure independent of the raw-handle ratchet:
collect_modules.rs reached 2009 lines against the 2000-line ceiling. The
three small pure helpers at the top move to a sibling module and are
re-imported by name, the recipe check_file_size.sh itself prints. They
share no state with the collection walk, so this is behaviour-free.
